requirements.txt
asn1crypto==0.24.0
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
cryptography==2.3
diff-match-patch==20121119
Django==1.9.8
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.0.1
django-pure-pagination==0.3.0
django-simple-captcha==0.4.6
enum34==1.1.6
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.9.2
idna==2.7
ipaddress==1.0.22
jdcal==1.4
MySQL-python==1.2.5
odfpy==1.3.6
openpyxl==2.5.4
Pillow==5.2.0
pycparser==2.18
pycryptodome==3.6.6
PyMySQL==0.9.2
PyYAML==3.13
requests==2.19.1
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.23
xlrd==1.1.0
xlwt==1.3.0
